Update buffer size calculation from Short to Float.
This change adjusts the buffer size calculation to use `Float.SIZE_BYTES` instead of `Short.SIZE_BYTES`. It ensures that the buffer size accurately reflects the data type being used, preventing possible memory allocation issues.

+++ core/src/commonMain/kotlin/com/littlekt/graphics/WebgpuExt.kt
@@ -30,7 +30,7 @@ fun Device.createGPUFloatBuffer(label: String, data: FloatArray, usages: Set<Buf
     val buffer = createBuffer(
         BufferDescriptor(
             label = label,
-            size = data.size.toLong() * Short.SIZE_BYTES,
+            size = data.size.toLong() * Float.SIZE_BYTES,
             usage = usages,
             mappedAtCreation = true
         )
